HID: i2c-hid: Retry PWR_ON command on resume to wake Weida devices
In 7d6f065de37c ("HID: i2c-hid: Use address probe to wake on resume"),
the retry of power on commands for the purpose of waking devices was
replaced with a generic bus probe, aligning boot and resume paths.

This caused a regression for Weida devices on resume from suspend, as
these devices are apparently responsive enough to ack their address and
pass the bus probe yet not responsive enough to succeed on their first
power command. It is unclear if they just need any command to arrive and
fail or if they strictly require a duplicate power-on command.

Re-introduce retry of power-on commands in the resume path with a more
elaborate description to avoid future regressions.

Fixes: 7d6f065de37c ("HID: i2c-hid: Use address probe to wake on resume")
Closes: https://bugzilla.kernel.org/show_bug.cgi?id=219440
Cc: stable@vger.kernel.org