HID: i2c-hid: Delay I2C resume wakeup for Weida devices as well
In 7d6f065de37c ("HID: i2c-hid: Use address probe to wake on resume"),
the retry of all power on commands was replaced with a generic bus
probe, aligning initial boot-up and resume paths.

This caused a regression for Weida devices on resume from suspend, as
these devices are apparently responsive enough to ack their address and
pass the bus probe yet not responsive enough to succeed on their first
power command.

Apply the new DELAY_WAKEUP_AFTE_RESUME quirk to Weida to give it more
time to settle.

Fixes: 7d6f065de37c ("HID: i2c-hid: Use address probe to wake on resume")
Closes: https://bugzilla.kernel.org/show_bug.cgi?id=219440
Cc: stable@vger.kernel.org