HID: i2c-hid: Revert to using power commands for wake-up on resume
7d6f065de37c ("HID: i2c-hid: Use address probe to wake on resume")
replaced the retry of power commands with the same bus probe we use on
boot which accounts for timing requirements for retry.

This made at least one Weida device very unhappy, as the bus probe
despite being successful somehow causes the following power command to
fail so hard that the device never lets go of the bus. This means that
even retries of the power command would fail on a timeout as the bus
remains busy.

Remove the bus probe on resume and instead reintroduce retry on of the
power command for wake-up purposes while respecting the new retry
timings.

Fixes: 7d6f065de37c ("HID: i2c-hid: Use address probe to wake on resume")
Closes: https://bugzilla.kernel.org/show_bug.cgi?id=219440
Cc: stable@vger.kernel.org
Signed-off-by: Kenny Levinsen <kl@kl.wtf>